Moved admin functions related to key operations to `common`
