+++ employee-vci/src/admin/routes.ts
@@ -4,8 +4,7 @@ import { koaBody } from "koa-body";
 import Router from "koa-router";
 
 import routesHandler from "./routesHandler.js";
-import { handleNotSuccessResult } from "../../../common/src/routerCommon.js";
-import keys from "../../../common/src/keys.js";
+import adminHandler from "../../../common/src/routes/admin/handler.js";
 
 const basicAuthOpts = () => {
   const name = process.env.BASIC_AUTH_USERNAME || "";
@@ -21,20 +20,7 @@ const init = () => {
     auth(basicAuthOpts()),
     koaBody(),
     async (ctx: Koa.Context) => {
-      if (!ctx.request.body) {
-        ctx.body = { status: "error", message: "Invalid data received!" };
-        ctx.status = 400;
-        return;
-      }
-      const kid = ctx.request.body.kid;
-      const curve = ctx.request.body.curve || "P-256";
-      const result = await keys.genKey(kid, curve);
-      if (result.ok) {
-        ctx.body = { status: "success", message: "Data created successfully!" };
-        ctx.status = 201;
-      } else {
-        handleNotSuccessResult(result.error, ctx);
-      }
+        await adminHandler.handleNewKey(ctx)
     },
   );
   router.post(
@@ -42,15 +28,7 @@ const init = () => {
     auth(basicAuthOpts()),
     koaBody(),
     async (ctx: Koa.Context) => {
-      // await routesHandler.handleRevokeKey(ctx);
-      const { kid } = ctx.params;
-      const result = await keys.revokeKey(kid);
-      if (result.ok) {
-        ctx.body = { status: "success", message: "Data revoked successfully!" };
-        ctx.status = 200;
-      } else {
-        handleNotSuccessResult(result.error, ctx);
-      }
+      await adminHandler.handleRevokeKey(ctx);
     },
   );
 
@@ -59,16 +37,7 @@ const init = () => {
     auth(basicAuthOpts()),
     koaBody(),
     async (ctx: Koa.Context) => {
-      // await routesHandler.handleCsr(ctx);
-      const { kid } = ctx.params;
-      const subject = ctx.request.body.subject;
-      const result = await keys.createCsr(kid, subject);
-      if (result.ok) {
-        ctx.body = { status: "success", payload: result.payload };
-        ctx.status = 200;
-      } else {
-        handleNotSuccessResult(result.error, ctx);
-      }
+      await adminHandler.handleCsr(ctx);
     },
   );
 
@@ -77,16 +46,7 @@ const init = () => {
     auth(basicAuthOpts()),
     koaBody(),
     async (ctx: Koa.Context) => {
-      // await routesHandler.handleSignSelfCert(ctx);
-      const { kid } = ctx.params;
-      const csr = ctx.request.body.csr;
-      const result = await keys.createSelfCert(kid, csr);
-      if (result.ok) {
-        ctx.body = { status: "success", payload: result.payload };
-        ctx.status = 200;
-      } else {
-        handleNotSuccessResult(result.error, ctx);
-      }
+      await adminHandler.handleSignSelfCert(ctx);
     },
   );
 
@@ -95,19 +55,7 @@ const init = () => {
     auth(basicAuthOpts()),
     koaBody(),
     async (ctx: Koa.Context) => {
-      // await routesHandler.handleRegisterCert(ctx);
-      const { kid } = ctx.params;
-      const certificates = ctx.request.body.certificates || [];
-      const result = await keys.registerCert(kid, certificates);
-      if (result.ok) {
-        ctx.body = {
-          status: "success",
-          message: "certificate registration succeeded",
-        };
-        ctx.status = 200;
-      } else {
-        handleNotSuccessResult(result.error, ctx);
-      }
+      await adminHandler.handleRegisterCert(ctx);
     },
   );
 
@@ -115,15 +63,7 @@ const init = () => {
     "/admin/keys/:kid",
     auth(basicAuthOpts()),
     async (ctx: Koa.Context) => {
-      // await routesHandler.handleGetKey(ctx);
-      const { kid } = ctx.params;
-      const result = await keys.getKey(kid);
-      if (result.ok) {
-        ctx.body = { status: "success", payload: result.payload };
-        ctx.status = 200;
-      } else {
-        handleNotSuccessResult(result.error, ctx);
-      }
+      await adminHandler.handleGetKey(ctx);
     },
   );
